ScanVideoS3Rekognition/requirements.txt
-e .
pytest
aws-cdk.aws_s3
aws-cdk.aws_s3_notifications
aws-cdk.aws_sns
aws-cdk.aws_lambda
aws-cdk.aws_dynamodb
aws-cdk.aws_lambda_event_sources
aws-cdk.core
aws-cdk.aws_sns_subscriptions
